stream_processor/requirements.txt
kafka-python
pandas
scikit-learn
requests
python-dotenv